RavenDB-18201 - change to using ProxyCommand
