+++ src/Raven.Server/Documents/Handlers/Processors/Attachments/AbstractAttachmentHandlerProcessorForGetAttachment.cs
@@ -26,7 +26,7 @@ namespace Raven.Server.Documents.Handlers.Processors.Attachments
             _isDocument = isDocument;
         }
 
-        protected abstract ValueTask<AttachmentResult> GetAttachmentAsync(TOperationContext context, string documentId, string name, AttachmentType type, string changeVector, CancellationToken token);
+        protected abstract ValueTask GetAttachmentAsync(TOperationContext context, string documentId, string name, AttachmentType type, string changeVector, CancellationToken token);
 
         protected abstract RavenTransaction OpenReadTransaction(TOperationContext context);
 
@@ -54,48 +54,7 @@ namespace Raven.Server.Documents.Handlers.Processors.Attachments
                         throw new ArgumentException("The 'ChangeVector' field in the body request is mandatory");
                 }
                 
-                var attachment = await GetAttachmentAsync(context, documentId, name, type, changeVector, token.Token);
-                if (attachment == null)
-                    return;
-
-                try
-                {
-                    var fileName = Path.GetFileName(attachment.Details.Name);
-                    fileName = Uri.EscapeDataString(fileName);
-                    HttpContext.Response.Headers[Constants.Headers.ContentDisposition] = $"attachment; filename=\"{fileName}\"; filename*=UTF-8''{fileName}";
-                }
-                catch (ArgumentException e)
-                {
-                    if (Logger.IsInfoEnabled)
-                        Logger.Info($"Skip Content-Disposition header because of not valid file name: {attachment.Details.Name}", e);
-                }
-                try
-                {
-                    HttpContext.Response.Headers[Constants.Headers.ContentType] = attachment.Details.ContentType;
-                }
-                catch (InvalidOperationException e)
-                {
-                    if (Logger.IsInfoEnabled)
-                        Logger.Info($"Skip Content-Type header because of not valid content type: {attachment.Details.ContentType}", e);
-                    if (HttpContext.Response.Headers.ContainsKey(Constants.Headers.ContentType))
-                        HttpContext.Response.Headers.Remove(Constants.Headers.ContentType);
-                }
-                HttpContext.Response.Headers[Constants.Headers.AttachmentHash] = attachment.Details.Hash;
-                HttpContext.Response.Headers[Constants.Headers.AttachmentSize] = attachment.Details.Size.ToString();
-                HttpContext.Response.Headers[Constants.Headers.Etag] = $"\"{attachment.Details.ChangeVector}\"";
-
-                using (context.GetMemoryBuffer(out var buffer))
-                await using (var stream = attachment.Stream)
-                {
-                    var responseStream = RequestHandler.ResponseBodyStream();
-                    var count = stream.Read(buffer.Memory.Memory.Span); // can never wait, so no need for async
-                    while (count > 0)
-                    {
-                        await responseStream.WriteAsync(buffer.Memory.Memory.Slice(0, count), token.Token);
-                        // we know that this can never wait, so no need to do async i/o here
-                        count = stream.Read(buffer.Memory.Memory.Span);
-                    }
-                }
+                await GetAttachmentAsync(context, documentId, name, type, changeVector, token.Token);
             }
         }
     }
